DLS-3690: Fixing SIDL compilation warnings (#104)

+++ app/uk/gov/hmrc/softdrinksindustrylevystub/services/SdilNumberTransformer.scala
@@ -16,7 +16,6 @@
 
 package uk.gov.hmrc.softdrinksindustrylevystub.services
 
-import org.scalacheck._
 import uk.gov.hmrc.smartstub._
 import uk.gov.hmrc.domain.Modulus23Check
 
@@ -44,8 +43,8 @@ object SdilNumberTransformer {
   def showTable(num: Int): Unit = {
     for {
       i       <- 1 to num
-      utr     <- tolerantUtr.get(i)
-      sdilRef <- sdilRefEnum.get(i)
+      utr     <- tolerantUtr.get(i.toLong)
+      sdilRef <- sdilRefEnum.get(i.toLong)
     } yield {
       println(s"| $utr | $sdilRef |")
     }
